requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.2
black==20.8b1
boto3==1.16.53
botocore==1.19.53
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
Collectfast==2.2.0
coverage==5.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.10
django-allauth==0.42.0
django-anymail==8.0
django-cors-headers==3.5.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
django-environ==0.4.5
django-extensions==3.0.9
django-model-utils==4.0.0
django-redis==4.12.1
django-storages==1.10.1
django-stubs==1.5.0
djangorestframework==3.11.1
docutils==0.16
factory-boy==3.0.1
Faker==4.1.3
filelock==3.0.12
flake8==3.8.3
flake8-isort==4.0.0
gunicorn==20.0.4
hiredis==1.1.0
identify==1.5.5
idna==2.10
imagesize==1.2.0
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.5.3
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
livereload==2.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
pre-commit==2.7.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.0.2
pytest-django==3.10.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.4
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinx-autobuild==2020.9.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.14.2
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
